requirements.txt
pandas
openpyxl
numpy
python-dotenv
google-genai
faiss-cpu
duckdb